UserApplication

+++ src/main/java/com/example/userLogin/webApi/controllers/UsersController.java
@@ -38,62 +38,97 @@ public class UsersController {
 	
 	private UserService userService;
 	
+	//********************************** KULLANICI KAYIT ************************************
+	
 	@PostMapping("/createUser")
 	@ResponseStatus(code = HttpStatus.CREATED)
 	public void add(@RequestBody @Valid CreateUserRequest createUserRequest) {
 		this.userService.add(createUserRequest);
 	}
 	
+	
+	//********************************** KULLANICI SİLME ************************************
+	
 	@DeleteMapping("/deleteUser/{id}")
 	public void delete(@PathVariable int id) {
 		this.userService.delete(id);
 	}
+	
+	
+	//********************************** KULLANICI GÜNCELLEME ************************************
 	
 	@PutMapping("/updateUser")
 	public void update(@RequestBody @Valid UpdateUserRequest updateUserRequest) {
 		this.userService.update(updateUserRequest);
 	}
 	
+	
+	//********************************** BELİRTİLEN ID NUMARASINA SAHİP KULLANICIYI GETİRME ************************************
+	
 	@GetMapping("/getByIdUser/{id}")
 	public GetByIdUserResponse getByIdUserResponse(@PathVariable int id) {
 		return this.userService.getById(id);
 	}
 	
+	
+	//********************************** BELİRTİLEN KULLANICI ADINA SAHİP KULLANICIYI GETİRME ************************************
+
 	@GetMapping("/getByUsername/{username}")
 	public GetByUsernameResponse getByUsernameResponse(@PathVariable @Size(min = 3,max = 10) String username) {
 		return this.userService.getByUsername(username);
 	}
+	
+	
+	//********************************** BELİRTİLEN AD VE SOYADA SAHİP KULLANICIYI GETİRME ************************************
 	
 	@GetMapping("/getByNameAndSurname/{name}/{surname}")
 	public GetByNameAndSurnameResponse getByNameAndSurnameResponse(@PathVariable String name,@PathVariable String surname) {
 		return this.userService.getByNameAndSurname(name, surname);
 	}
 	
+	
+	//********************************** BELİRTİLEN TELEFON NUMARASINA SAHİP KULLANICIYI GETİRME ************************************
+
 	@GetMapping("/getByPhoneNumber/{phoneNumber}")
 	public GetByPhoneNumberResponse getByPhoneNumberResponse(@PathVariable @Size(min = 11,max = 11) String phoneNumber){
 		return this.userService.getByPhoneNumberResponse(phoneNumber);
 	}
 	
+	
+	//********************************** BÜTÜN KULLANICILARI GETİRME ************************************
+
 	@GetMapping("/getAllUsers")
 	public List<GetAllUsersResponse> getAll(){
 		return this.userService.getAll();
 	}
 	
+	
+	//********************************** BELİRTİLEN CİNSİYETTEKİ KULLANICILARI GETİRME ************************************
+
 	@GetMapping("/getAllByGenderId/{id}")
 	public List<GetAllByGenderIdResponse> getAllByGenderId(@PathVariable byte id){
 		return this.userService.getAllByGenderId(id);
 	}
 	
+	
+	//********************************** BELİRTİLEN ÜLKELERDEKİ KULLANICILARI GETİRME ************************************
+
 	@GetMapping("/getAllByCountryIdIn/{countriesId}")
 	public List<GetAllByCountryIdInResponse> getAllByCountryIdIn(@PathVariable List<Integer> countriesId){
 		return this.userService.getAllByCountryIdIn(countriesId);
 	}
 	
+	
+	//********************************** BELİRTİLEN ÜLKEDEKİ SEÇİLEN ŞEHİRLERDEKİ KULLANICILARI GETİRME ************************************
+
 	@GetMapping("/getAllByCountryIdAndCityIdIn/{countryId}/{citiesId}")
-	public List<GetAllByCountryIdAndCityIdInResponse> getAllByCountryIdAndCityIdIn(@PathVariable byte countryId,@PathVariable List<Integer> citiesId){
+	public List<GetAllByCountryIdAndCityIdInResponse> getAllByCountryIdAndCityIdIn(@PathVariable int countryId,@PathVariable List<Integer> citiesId){
 		return this.userService.getAllByCountryIdAndCityIdIn(countryId, citiesId);
 	}
 	
+	
+	//********************************** AKTİF OLAN KULLANICILARI GETİRME ************************************
+
 	@GetMapping("/getAllByActiveTrue")
 	public List<GetAllByActiveTrueResponse> getAllByActiveTrueResponses(){
 		return this.userService.getAllByActiveTrueResponse();
